Add authorization

+++ src/app/components/profile/profile.component.ts
@@ -6,6 +6,7 @@ import { UserService } from 'src/app/services/user.service';
 import { FormBuilder } from '@angular/forms';
 import { NavigationService } from 'src/app/services/navigation.service';
 import { FollowService } from 'src/app/services/follow.service';
+import { AccountService } from 'src/app/services/account.service';
 
 @Component({
   selector: 'app-profile',
@@ -36,19 +37,36 @@ export class ProfileComponent {
     groupChatMessages: [],
     notifications: []
   };
+  unathorizedMessage: string;
+  isUnathorizedFollowClicked: boolean = false;
+  toggleUnathorizedFollowClick() {
+    if (!this.accountService.isAuthorized()) {
+      if (this.isUnathorizedFollowClicked) {
+        document.body.style.overflow = 'auto';
+        document.body.style.marginRight = '0'
+        this.isUnathorizedFollowClicked = false;
+      }
+      else {
+        document.body.style.marginRight = '15px'
+        document.body.style.overflow = 'hidden';
+        this.isUnathorizedFollowClicked = true
+      }
+    }
+  }
   showUserPosts: boolean;
   showUserComments: boolean;
   isEditing: boolean = false;
   isFollowed: boolean = false;
   followers: IFollow[] = [];
   following: IFollow[] = [];
-  constructor(private fb: FormBuilder, private navigation: NavigationService, private route: ActivatedRoute, private userService: UserService, private followService: FollowService) { }
+  constructor(private accountService: AccountService, private fb: FormBuilder, private navigation: NavigationService, private route: ActivatedRoute, private userService: UserService, private followService: FollowService) { }
   ngOnInit() {
     this.route.params.subscribe(params => {
       this.userService.getByUserName(params['userName']).subscribe(res => {
         this.accountOwnerId = res.id
-        this.currentUserId = this.userService.getCurrentUserId()
-        if (this.userService.getCurrentUserId() != undefined) [
+        this.unathorizedMessage = `Start reading ${res.displayUsername} to see what he or she shares.`
+        this.currentUserId = this.accountService.getCurrentUserId()
+        if (this.accountService.getCurrentUserId() != undefined) [
           this.userService.getById(this.currentUserId).subscribe(res => {
             this.currentUser = res
             this.userChangesForm.get("displayUsername")?.setValue(this.currentUser.displayUsername);
@@ -111,26 +129,29 @@ export class ProfileComponent {
     }
   }
   toggleFollow(): void {
-    if (this.isFollowed) {
-      const follow = this.followers.find(follow => follow.followerId == this.currentUserId);
-      if (follow) {
-        this.followService.delete(follow.id).subscribe(res => {
+    this.toggleUnathorizedFollowClick();
+    if (this.accountService.isAuthorized()) {
+      if (this.isFollowed) {
+        const follow = this.followers.find(follow => follow.followerId == this.currentUserId);
+        if (follow) {
+          this.followService.delete(follow.id).subscribe(res => {
+            this.followService.getByFollowedUserId(this.accountOwnerId).subscribe(res => { this.followers = res })
+          });
+        }
+      }
+      else {
+        const newFollow: IFollow = {
+          id: 0,
+          followerId: this.currentUserId,
+          followedUserId: this.accountOwnerId,
+        };
+        this.followService.create(newFollow).subscribe(res => {
+          this.followers.push(res);
           this.followService.getByFollowedUserId(this.accountOwnerId).subscribe(res => { this.followers = res })
         });
       }
+      this.isFollowed = !this.isFollowed;
     }
-    else {
-      const newFollow: IFollow = {
-        id: 0,
-        followerId: this.currentUserId,
-        followedUserId: this.accountOwnerId,
-      };
-      this.followService.create(newFollow).subscribe(res => {
-        this.followers.push(res);
-        this.followService.getByFollowedUserId(this.accountOwnerId).subscribe(res => { this.followers = res })
-      });
-    }
-    this.isFollowed = !this.isFollowed;
   }
   showPosts(): void {
     this.showUserPosts = true;
